+++ pkg/kics/resolver_sink.go
@@ -5,6 +5,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/Checkmarx/kics/pkg/minified"
 	"sort"
 
 	sentryReport "github.com/Checkmarx/kics/internal/sentry"
@@ -29,8 +30,9 @@ func (s *Service) resolverSink(ctx context.Context, filename, scanID string, ope
 		s.Tracker.TrackFileFound()
 		countLines := bytes.Count(rfile.Content, []byte{'\n'}) + 1
 		s.Tracker.TrackFileFoundCountLines(countLines)
-		//burro aqui sff
-		documents, err := s.Parser.Parse(rfile.FileName, rfile.Content, openAPIResolveReferences, false)
+
+		isMinified := minified.IsMinified(rfile.FileName, rfile.Content)
+		documents, err := s.Parser.Parse(rfile.FileName, rfile.Content, openAPIResolveReferences, isMinified)
 		if err != nil {
 			if documents.Kind == "break" {
 				return []string{}, nil
@@ -69,6 +71,7 @@ func (s *Service) resolverSink(ctx context.Context, filename, scanID string, ope
 				LinesIgnore:       documents.IgnoreLines,
 				ResolvedFiles:     documents.ResolvedFiles,
 				LinesOriginalData: utils.SplitLines(string(rfile.OriginalData)),
+				IsMinified:        documents.IsMinified,
 			}
 			s.saveToFile(ctx, &file)
 		}
